static_casting inside of Projectile::Update();

+++ Source/Projectile.cpp
@@ -6,7 +6,7 @@ Projectile::Projectile(Vector2 pos, Vector2 vel) : Entity(pos, vel, {10, 30}) {}
 void Projectile::Update() {
 	Entity::Update(); 
 
-	if (position.y < -size.y || position.y > GetScreenHeight()) {
+	if (position.y < -size.y || position.y > static_cast<float>(GetScreenHeight())) {
 		active = false; 
 	}
 }
